myproject/core/api/views/auth/logout_api.py
"""
Logout API endpoint.

Handles user logout and token blacklisting.
"""

from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status, permissions
from rest_framework_simplejwt.tokens import RefreshToken

class LogoutAPIView(APIView):
    """
    API endpoint for user logout.

    Blacklists the refresh token to prevent reuse.
    POST /api/auth/logout/
    """
    permission_classes = [permissions.IsAuthenticated]

    def post(self, request):
        """
        Log out user by blacklisting refresh token.

        Expected data:
            {"refresh": "token_value"}
        """
        # TODO: Implement logout with token blacklisting
        # refresh_token = request.data.get('refresh')
        # if refresh_token:
        #     try:
        #         token = RefreshToken(refresh_token)
        #         token.blacklist()
        #         return Response({'detail': 'Successfully logged out'})
        #     except Exception as e:
        #         return Response(
        #             {'error': 'Invalid token'},
        #             status=status.HTTP_400_BAD_REQUEST
        #         )
        # return Response(
        #     {'error': 'Refresh token required'},
        #     status=status.HTTP_400_BAD_REQUEST
        # )

        return Response(
            {'detail': 'Not implemented'},
            status=status.HTTP_501_NOT_IMPLEMENTED
        )
