myproject/requirements.txt
# Core Django
Django>=5.0,<6.0

# REST Framework
djangorestframework>=3.15

# Environment configuration
python-decouple>=3.8

# PostgreSQL adapter
psycopg2-binary>=2.9

# Image processing
Pillow>=10.0

# CORS headers
django-cors-headers>=4.3

# JWT authentication
djangorestframework-simplejwt>=5.3

# API documentation
drf-yasg>=1.21

# Filtering
django-filter>=24.2
